feat: di api

+++ widget_compose/lib/entities/product.dart
@@ -3,7 +3,12 @@ class ProductToDisplay {
   late final String category;
   late final String imageUrl;
   late final String name;
-  late final int price;
+  late final double price;
 
-  ProductToDisplay({required this.id, required this.category, required this.imageUrl, required this.name, required this.price});
+  ProductToDisplay(
+      {required this.id,
+      required this.category,
+      required this.imageUrl,
+      required this.name,
+      required this.price});
 }
